Move ZigbeeOTA setDataDir, readLocalFile, getFirmwareFile to ota/common.ts
